requirements.txt
aiohttp==3.5.4
amqp==2.4.0
async-timeout==3.0.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
idna==2.8
kombu==4.2.2.post1
multidict==4.5.2
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
vine==1.2.0
yarl==1.3.0
